local_superstore/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
altair==4.0.0
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.2
apprise==1.4.0
arrow==1.2.3
asgi-lifespan==2.1.0
asn1crypto==1.5.1
astor==0.8.1
async-timeout==4.0.3
asyncpg==0.27.0
attrs==23.1.0
Authlib==0.15.5
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.2.2
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
clikit==0.6.2
cloudpickle==2.0.0
cmdstanpy==0.9.5
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
convertdate==2.4.0
coolname==2.2.0
crashtest==0.3.1
croniter==1.3.15
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.2
dateparser==1.1.8
debugpy==1.5.1
decorator==5.1.0
docker==6.1.2
duckdb==0.8.1
entrypoints==0.3
ephem==4.1.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Faker==19.6.2
fastapi==0.95.2
favicon==0.7.0
filelock==3.12.3
Flask==2.0.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.5.0
geographiclib==2.0
geopy==2.4.0
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==2.0.2
griffe==0.29.0
gw_dsl_parser==0.1.6a1
h11==0.14.0
h2==4.1.0
holidays==0.32
hpack==4.0.0
htbuilder==0.6.2
httpcore==0.17.2
httpstan==4.10.1
httpx==0.24.1
hydralit-components==1.0.10
hyperframe==6.0.1
idna==3.3
importlib-metadata==6.6.0
importlib-resources==6.0.1
ipykernel==6.5.0
ipython==7.29.0
ipywidgets==8.1.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kubernetes==26.1.0
LunarCalendar==0.0.9
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.7.3
matplotlib-inline==0.1.3
mdurl==0.1.2
millify==0.1.1
more-itertools==10.1.0
MouseInfo==0.1.3
multidict==6.0.4
nest-asyncio==1.5.1
numerize==0.12
numpy==1.24.3
oauthlib==3.1.1
openpyxl==3.1.2
orjson==3.8.14
oscrypto==1.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.2
pastel==0.2.1
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
Plost==0.2.5
plotly==5.16.1
prefect==2.10.11
prompt-toolkit==3.0.22
prophet==1.1.4
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py4j==0.10.9.7
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.8
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.15.1
pygwalker==0.3.7
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
pysimdjson==5.0.2
pyspark==3.4.1
pystan==3.7.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytweening==1.0.7
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
readchar==4.0.5
regex==2023.5.5
requests==2.26.0
requests-oauthlib==1.3.0
rich==13.3.5
rsa==4.8
rubicon-objc==0.4.6
schedules==1.3
semver==3.0.1
setuptools-git==1.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==3.2.0
snowflake-snowpark-python==1.7.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.15
sqlglot==18.7.0
sqlparse==0.4.4
st-annotated-text==4.0.1
starlette==0.27.0
streamlit==1.12.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.1
tqdm==4.66.1
traitlets==5.1.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.7
uvicorn==0.22.0
validators==0.22.0
wasmtime==12.0.0
wcwidth==0.2.5
webargs==8.3.0
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.0.2
widgetsnbextension==4.0.9
wordcloud==1.9.2
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.15.0
